lpbin: add get-audio-tags signal for upper layer
Changed signals of audio and video streams are added
to notify upper layer.
To get the tag information of audio and video streams,
get-audio-tags and get_video_tags signals are added.
Also, To get a pad instance of audio and video streams,
get_video_pad and get_audio_pad signals are added.

:Issues Addressed:
[GF-41484] add get-audio-tags signal at lpbin

Change-Id: I8ceddeada5ce9a6f5a0627da1c87f322078532ae
Reviewed-on: https://gpro.palm.com/13206
Tested-by: HoonHee Lee <hoonhee.lee@lge.com>
Reviewed-by: Hyunuk Seo <hyunuk.seo@lge.com>
Tested-by: Hyunuk Seo <hyunuk.seo@lge.com>
Reviewed-by: Hyunrae Cho <hr.cho@lge.com>
Reviewed-by: Jeongseok Kim <jeongseok.kim@lge.com>