lpsink: Add video and audio channel information
A/V channel information for hardware decoders can be acquired by 
only the upper layer's logic. 

Each bit of channel properties has meaning of port number.
For example, if the zero bit is set as 1, the channel 0 is required.
If first and second bits are set as 1, both channels, 0 and 1, are 
required. 

Dual streaming is only exceptional case because the lower layer will
control the channel information.

:Issues Addressed:
[GF-57923] Dynamic resource allocation for generic pipeline

Change-Id: I33bba4506458a146d534688e3a31ac1f4b6c1349
Reviewed-on: https://gpro.palm.com/24902
Tested-by: HoonHee Lee <hoonhee.lee@lge.com>
Reviewed-by: Jeongseok Kim <jeongseok.kim@lge.com>
Tested-by: Jeongseok Kim <jeongseok.kim@lge.com>
Reviewed-on: https://gpro.palm.com/26927